backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
openai
python-dotenv
pydantic
